WI #2699 Use separate namespaces for each command
